Pulled GuiTab out of GuiWindow.
